Test case that illustrates the bug
The default implementation of `wNoThunks` translates its argument to
a generic representation, _then forces that generic representation to
WHNF_, and then calls `gwNoThunks`. This is wrong; given a datatype with
a single constructor with a single field

```
data T = MkT Int
```

the generic representation of `MkT x` is `M1 (M1 (K1 x))`, where `M1`
and `K1` are both newtypes. This means that forcing the representation
to WHNF means we would force the `Int` to WHNF, which is incorrect:
calling `noThunks` should never force any values to WHNF that aren't
already.

Admittedly, this is an edge case, but I think the forcing to WHNF
is simply not required: the implementation of `gwNoThunks` does not rely
on it.

This commit does not fix the problem, only adds the test case that
illustrates it.